# Group2_CompRepair/Group2_CompRepair/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Group2_CompRepair, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/ComputerParts": {
      "get": {
        "tags": [
          "ComputerParts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ComputerPart"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ComputerParts"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ComputerPart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ComputerPart"
            }
          }
        }
      }
    },
    "/api/ComputerParts/{id}": {
      "get": {
        "tags": [
          "ComputerParts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ComputerPart"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ComputerParts"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ComputerPart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ComputerParts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Customer"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customers"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        }
      }
    },
    "/api/Customers/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Customers"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employees": {
      "get": {
        "tags": [
          "Employees"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        }
      }
    },
    "/api/Employees/{id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employees"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      }
    },
    "/api/Orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Softwares": {
      "get": {
        "tags": [
          "Softwares"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Software"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Softwares"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Software"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Software"
            }
          }
        }
      }
    },
    "/api/Softwares/{id}": {
      "get": {
        "tags": [
          "Softwares"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Software"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Softwares"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Software"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Softwares"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "ComputerPart": {
      "type": "object",
      "properties": {
        "computerPartsId": {
          "format": "int32",
          "type": "integer"
        },
        "partName": {
          "type": "string"
        },
        "partDescription": {
          "type": "string"
        },
        "partsInStock": {
          "format": "int32",
          "type": "integer"
        },
        "partType": {
          "type": "string"
        },
        "partPrice": {
          "format": "double",
          "type": "number"
        },
        "partsOrdersLinkings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PartsOrdersLinking"
          }
        }
      }
    },
    "Customer": {
      "type": "object",
      "properties": {
        "customerId": {
          "format": "int32",
          "type": "integer"
        },
        "customerFirstname": {
          "type": "string"
        },
        "customerLastname": {
          "type": "string"
        },
        "customerPhone": {
          "type": "string"
        },
        "customerEmail": {
          "type": "string"
        },
        "customerAddress": {
          "type": "string"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        }
      }
    },
    "Employee": {
      "type": "object",
      "properties": {
        "employeeId": {
          "format": "int32",
          "type": "integer"
        },
        "employeeFirstname": {
          "type": "string"
        },
        "employeeLastname": {
          "type": "string"
        },
        "employeePhone": {
          "type": "string"
        },
        "employeeEmail": {
          "type": "string"
        },
        "employeeAddress": {
          "type": "string"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "customerId": {
          "format": "int32",
          "type": "integer"
        },
        "employeeId": {
          "format": "int32",
          "type": "integer"
        },
        "orderTotal": {
          "format": "double",
          "type": "number"
        },
        "orderDate": {
          "format": "date-time",
          "type": "string"
        },
        "customer": {
          "$ref": "#/definitions/Customer"
        },
        "employee": {
          "$ref": "#/definitions/Employee"
        },
        "partsOrdersLinkings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PartsOrdersLinking"
          }
        },
        "softwareOrdersLinkings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SoftwareOrdersLinking"
          }
        }
      }
    },
    "PartsOrdersLinking": {
      "type": "object",
      "properties": {
        "partsOrdersId": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "computerPartsId": {
          "format": "int32",
          "type": "integer"
        },
        "qtyOfParts": {
          "format": "int32",
          "type": "integer"
        },
        "computerParts": {
          "$ref": "#/definitions/ComputerPart"
        },
        "order": {
          "$ref": "#/definitions/Order"
        }
      }
    },
    "Software": {
      "type": "object",
      "properties": {
        "softwareId": {
          "format": "int32",
          "type": "integer"
        },
        "softwareName": {
          "type": "string"
        },
        "softwareDescription": {
          "type": "string"
        },
        "licensesInStock": {
          "format": "int32",
          "type": "integer"
        },
        "softwareType": {
          "type": "string"
        },
        "softwarePrice": {
          "format": "double",
          "type": "number"
        },
        "softwareOrdersLinkings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SoftwareOrdersLinking"
          }
        }
      }
    },
    "SoftwareOrdersLinking": {
      "type": "object",
      "properties": {
        "softwareOrdersId": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "softwareId": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "$ref": "#/definitions/Order"
        },
        "software": {
          "$ref": "#/definitions/Software"
        }
      }
    }
  }
}